Dropped the Signature proto and replaced it with SignalAgglomUpdates.proto.
Signature, formerly known as encrypted Unknown 6 payload was a name we came up with earlier. A few weeks has passed and all types and most fields of this protocol is known now. As we all know it acts as some kind of agglomerated package of all changed device and sensor information. That's the reason why SignalAgglomUpdates.proto was choosen here.

Besides this, all types and missing fields are available now. This protocol is dumped from the current 0.35 client. As noted in the protocol itself all types are confirmed except some int32 values that may be floats.

Have fun and prepare your MITM attacks to guess the missing values.

Your loved,
Buddy Team